Address additional review comments
