+++ packages/suite-base/src/panels/ThreeDeeRender/Interactions/constants.ts
@@ -28,24 +28,3 @@ export const HOVER_TOOLTIP_MAX_W = 620;
 
 /** Maximum tooltip height (px) – used for positioning calculations. Must match CSS maxHeight. */
 export const HOVER_TOOLTIP_MAX_H = 480;
-
-/**
- * Clamp a tooltip along a single axis so it stays inside the given bounds.
- * Prefers placing it after the cursor; flips before if there isn't space;
- * falls back to the nearest edge otherwise.
- */
-export function clampTooltipAxis(
-  cursor: number,
-  size: number,
-  boundsStart: number,
-  boundsEnd: number,
-  offset: number,
-): number {
-  if (boundsEnd - cursor >= size + offset) {
-    return cursor + offset;
-  }
-  if (cursor - boundsStart >= size + offset) {
-    return cursor - size - offset;
-  }
-  return Math.max(boundsStart, boundsEnd - size);
-}
